added authenticationa nd authorisation as well

+++ blog/views.py
@@ -1,3 +1,7 @@
+from django.contrib.auth import login
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from django.core.exceptions import PermissionDenied
 from django.shortcuts import get_object_or_404, redirect, render
 
 from .forms import PostForm
@@ -14,19 +18,25 @@ def post_detail(request, post_id):
     return render(request, "blog/post_detail.html", {"post": post})
 
 
+@login_required
 def post_create(request):
     if request.method == "POST":
         form = PostForm(request.POST)
         if form.is_valid():
-            post = form.save()
+            post = form.save(commit=False)
+            post.author = request.user
+            post.save()
             return redirect("post_detail", post_id=post.id)
     else:
         form = PostForm()
     return render(request, "blog/post_form.html", {"form": form})
 
 
+@login_required
 def post_update(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.author != request.user:
+        raise PermissionDenied
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
         if form.is_valid():
@@ -37,9 +47,24 @@ def post_update(request, post_id):
     return render(request, "blog/post_form.html", {"form": form, "post": post})
 
 
+@login_required
 def post_delete(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.author != request.user:
+        raise PermissionDenied
     if request.method == "POST":
         post.delete()
         return redirect("post_list")
     return render(request, "blog/post_confirm_delete.html", {"post": post})
+
+
+def signup(request):
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)
+            return redirect("post_list")
+    else:
+        form = UserCreationForm()
+    return render(request, "registration/signup.html", {"form": form})
